+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## 10.5.3
 
 ### Changed
-*  Restructured consent object that is being sent to the backend API for biometric kyc, enhanced kyc and enhanced document verification
+* Restructured consent object that is being sent to the backend API for biometric kyc, enhanced kyc and enhanced document verification
 
 ## 10.5.2